added networkx module and tests and abstracted out the mcqgraph into its own baseclass
