+++ .pylintrc
@@ -197,6 +197,9 @@ good-names=a,
            i,
            j,
            k,
+           e,
+           n,
+           r,
            ex,
            Run,
            _
